Centralize Firestore timestamp conversion utilities
Moved Firestore timestamp conversion logic to a new utils/firestoreTimestamps.ts file and updated usage in LiquidityRequestsCard and useUserVaultsSummaries to use these utilities. Also improved error logging in AcceptConfirm and refactored USDC mainnet token ID handling in tokens.ts for better maintainability.
