+++ utils/tokens.ts
@@ -1,5 +1,8 @@
 import { getActiveNetwork, type Network } from "@/utils/networks";
 
+// Implicit account for USDC on NEAR mainnet
+const USDC_MAINNET_ID = "17208628f84f5d6ad33f0da3bbbeb27ffcb398eac501a31bd6ad2011e36133a1";
+
 export type TokenConfig = {
   id: string; // NEP-141 contract account id
   symbol: string;
@@ -24,8 +27,8 @@ const TOKENS_BY_NETWORK: Partial<Record<Network, Record<string, TokenConfig>>> =
     // USDC token on NEAR mainnet (implicit account). See also:
     // - https://github.com/near/near-discovery-token-registry/blob/main/mainnet.json
     // - https://github.com/sudostake/sudostake_agent_near/blob/main/agent/src/token_registry.py
-    "17208628f84f5d6ad33f0da3bbbeb27ffcb398eac501a31bd6ad2011e36133a1": {
-      id: "17208628f84f5d6ad33f0da3bbbeb27ffcb398eac501a31bd6ad2011e36133a1",
+    [USDC_MAINNET_ID]: {
+      id: USDC_MAINNET_ID,
       symbol: "USDC",
       decimals: 6,
       name: "USD Coin",
